requirements.txt
bcrypt==4.2.0
mysql-connector-python==9.0.0
python-dotenv==1.0.1
reportlab==4.2.2
matplotlib==3.9.2
pandas==2.2.2
pytest==8.3.2